+++ release_notes.md
@@ -3,6 +3,7 @@
 ## 4.0.3
 
 -   APP-4245 - [Inputs] Added support for secure Redis (SSL and User/Pass)
+-   APP-4246 - [Playbook] Removed support for AOT execution
 
 ## 4.0.2